[feat] util에 시간 계산 함수 추가
