+++ static/js/util.js
@@ -10,4 +10,13 @@ const getCSRF = () => {
     if (parts.length === 2) return parts.pop().split(';').shift();
 }
 
-export { getGroupId, getCSRF }
+const calculateTimeDifference = (endTime) => {
+    const timeDifference = endTime - currentTime;
+    const days = Math.floor(timeDifference / (1000 * 60 * 60 * 24));
+    const hours = Math.floor((timeDifference % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
+    const minutes = Math.floor((timeDifference % (1000 * 60 * 60)) / (1000 * 60));
+  
+    return { days, hours, minutes };
+  }
+
+export { getGroupId, getCSRF, calculateTimeDifference }
